ICE lite
Prep for hosted.

ALSO:
There might still be something wrong with broadcast autodiscovery...  I've had like two failed connections out of maybe 60 in the past week or so.  I had assumed that this was due to allocated ports being out of "safe" range, but I fixed that so it must be something else.