socketaddr btree key
Use a socket address instead of u64 as the btreemap's key.  This moves the port into the key, and as a result means one token may map to multiple values.  So instead of using get we must iterate over a range of keys.

In the future I intend to assign keys differently so this was neccessary anyway.  The keys are probably going to be assigned via Turn Username somehow.