Increase recv, maintain send
Increase the size of the common input buffer, but keep the TCP send buffers the same size.

Technically the UDP handling reuses the input buffer for send, so this increases the max size for sending UDP, but I expect the OS limit on UDP to manage that front.  It's only the BuffWriter wrapping the TCP stream that I'm concerned about.

WebRTC clients should already be testing the MTU through the TURN server so hopefully this is fine.

I'm primarily doing this because the logs for stun.evan-brass.net show many closed TCP connections with "STUN message too large to fit in buffer".  I'm sorry for the kind folk who have tested my service.  I hope this improves its usefulness for you.