dyndns
Use nsupdate to map/unmap pid -> AAAA.
Had to run hosted as root so it has access to the local tsig keys, and is able to nsupdate to the local domain name.  The right way to do this would be to use a dedicated key and permit it to update the zone, but that's not what I did.