Begin rewrite
I've moved ICE dissolve into a separate bin, and TURN is going to be split into udp and tcp (with potential to support channel-number on tcp).
Also moving to have real subnets for the stateless mapping (at least a /96 for ipv4).  Also would like to have a real subnet for tcp (possibly a /64 or /72 with 1 byte of the port being used to construct the usize (u64) index of the slab allocator.

Also, hosted is gonna be split into a dtls-proxy separate from an sctp vpn since I want to remove all of the DNS crap.