Fuck'n checksums man
The checksum doesn't need to be valid during routing, however it is verified when eating the packet.  This wasn't a problem before because I wasn't encapsulating the SCTP packets inside UDP.  Yes, this means there is a UDP checksum, and an SCTP checksum on the same fucking dipshit of a packet.